Refactor API-server database initialization code.
Split database initialization and database migration code into two modules.

Prepares for #3130.
